+++ packages/framer-motion/src/projection/node/types.ts
@@ -37,6 +37,7 @@ export type LayoutEvents =
 export interface IProjectionNode<I = unknown> {
     id: number
     animationId: number
+    animationCommitId: number
     parent?: IProjectionNode
     relativeParent?: IProjectionNode
     root?: IProjectionNode
@@ -102,7 +103,10 @@ export interface IProjectionNode<I = unknown> {
     applyTransform(box: Box, transformOnly?: boolean): Box
     resolveTargetDelta(force?: boolean): void
     calcProjection(): void
-    getProjectionStyles(styleProp?: MotionStyle): MotionStyle | undefined
+    applyProjectionStyles(
+        targetStyle: CSSStyleDeclaration,
+        styleProp?: MotionStyle
+    ): void
     clearMeasurements(): void
     resetTree(): void