+++ packages/motion-dom/src/animation/types.ts
@@ -5,6 +5,8 @@ import { Driver } from "./drivers/types"
 import { KeyframeResolver } from "./keyframes/KeyframesResolver"
 import { WithRender } from "./keyframes/types"
 
+export type AnyResolvedKeyframe = string | number
+
 export interface ProgressTimeline {
     currentTime: null | { value: number }
 
@@ -12,7 +14,7 @@ export interface ProgressTimeline {
 }
 
 export interface ValueAnimationOptionsWithRenderContext<
-    V extends string | number = number
+    V extends AnyResolvedKeyframe = number
 > extends ValueAnimationOptions<V> {
     KeyframeResolver?: typeof KeyframeResolver
     motionValue?: MotionValue<V>
@@ -110,8 +112,9 @@ export interface KeyframeGenerator<V> {
     toString: () => string
 }
 
-export interface DOMValueAnimationOptions<V extends string | number = number>
-    extends ValueAnimationTransition<V> {
+export interface DOMValueAnimationOptions<
+    V extends AnyResolvedKeyframe = number
+> extends ValueAnimationTransition<V> {
     element: HTMLElement | SVGElement
     keyframes: ValueKeyframesDefinition
     name: string
@@ -119,7 +122,7 @@ export interface DOMValueAnimationOptions<V extends string | number = number>
     allowFlatten?: boolean
 }
 
-export interface ValueAnimationOptions<V extends string | number = number>
+export interface ValueAnimationOptions<V extends AnyResolvedKeyframe = number>
     extends ValueAnimationTransition {
     keyframes: V[]
     element?: any // TODO: Replace with VisualElement when moved into motion-dom
@@ -375,11 +378,13 @@ export interface AnimationOrchestrationOptions {
 
     /**
      * When using variants, children animations will start after this duration
-     * (in seconds). You can add the `transition` property to both the `Frame` and the `variant` directly. Adding it to the `variant` generally offers more flexibility, as it allows you to customize the delay per visual state.
+     * (in seconds). You can add the `transition` property to both the `motion.div` and the
+     * `variant` directly. Adding it to the `variant` generally offers more flexibility,
+     * as it allows you to customize the delay per visual state.
      *
      * @public
      */
-    delayChildren?: number
+    delayChildren?: number | DynamicOption<number>
 
     /**
      * When using variants, animations of child components can be staggered by this
@@ -391,7 +396,7 @@ export interface AnimationOrchestrationOptions {
      *
      * The calculated stagger delay will be added to `delayChildren`.
      *
-     * @public
+     * @deprecated - Use `delayChildren: stagger(interval)` instead.
      */
     staggerChildren?: number
 
@@ -401,7 +406,7 @@ export interface AnimationOrchestrationOptions {
      * A value of `1` staggers from the first to the last while `-1`
      * staggers from the last to the first.
      *
-     * @public
+     * @deprecated - Use `delayChildren: stagger(interval, { from: "last" })` instead.
      */
     staggerDirection?: number
 }
@@ -477,7 +482,7 @@ export type SVGTransitions = {
     [K in keyof Omit<SVGAttributes, "from">]: ValueTransition
 }
 
-export type VariableTransitions = {
+export interface VariableTransitions {
     [key: `--${string}`]: ValueTransition
 }
 
@@ -485,10 +490,11 @@ export type StyleTransitions = {
     [K in keyof CSSStyleDeclarationWithTransform]?: ValueTransition
 }
 
-export type ValueKeyframe<T extends string | number = string | number> = T
+export type ValueKeyframe<T extends AnyResolvedKeyframe = AnyResolvedKeyframe> =
+    T
 
 export type UnresolvedValueKeyframe<
-    T extends string | number = string | number
+    T extends AnyResolvedKeyframe = AnyResolvedKeyframe
 > = ValueKeyframe<T> | null
 
 export type ResolvedValueKeyframe = ValueKeyframe | ValueKeyframe[]
@@ -506,7 +512,7 @@ export type SVGKeyframesDefinition = {
     [K in keyof Omit<SVGAttributes, "from">]?: ValueKeyframesDefinition
 }
 
-export type VariableKeyframesDefinition = {
+export interface VariableKeyframesDefinition {
     [key: `--${string}`]: ValueKeyframesDefinition
 }
 
@@ -524,9 +530,19 @@ export type DOMKeyframesDefinition = StyleKeyframesDefinition &
     SVGForcedAttrKeyframesDefinition &
     VariableKeyframesDefinition
 
+export interface Target extends DOMKeyframesDefinition {}
+
+type CSSPropertyKeys = {
+    [K in keyof CSSStyleDeclaration as K extends string
+        ? CSSStyleDeclaration[K] extends AnyResolvedKeyframe
+            ? K
+            : never
+        : never]: CSSStyleDeclaration[K]
+}
+
 export interface CSSStyleDeclarationWithTransform
     extends Omit<
-        CSSStyleDeclaration,
+        CSSPropertyKeys,
         "direction" | "transition" | "x" | "y" | "z"
     > {
     x: number | string
@@ -546,47 +562,67 @@ export interface CSSStyleDeclarationWithTransform
     scaleZ: number
     skewX: number | string
     skewY: number | string
+    transformPerspective: number
 }
 
-export type Transition<V = any> = StyleTransitions &
+export type TransitionWithValueOverrides<V> = ValueAnimationTransition<V> &
+    StyleTransitions &
     SVGPathTransitions &
     SVGForcedAttrTransitions &
     SVGTransitions &
-    VariableTransitions &
-    ValueAnimationTransition<V> & {
+    VariableTransitions & {
         default?: ValueTransition
         layout?: ValueTransition
     }
 
+export type Transition<V = any> =
+    | ValueAnimationTransition<V>
+    | TransitionWithValueOverrides<V>
+
 export type DynamicOption<T> = (i: number, total: number) => T
 
-export interface AnimationOptions extends Omit<Transition, "delay"> {
+export type ValueAnimationWithDynamicDelay = Omit<
+    ValueAnimationTransition<any>,
+    "delay"
+> & {
     delay?: number | DynamicOption<number>
 }
 
+export type AnimationOptions =
+    | ValueAnimationWithDynamicDelay
+    | (ValueAnimationWithDynamicDelay &
+          StyleTransitions &
+          SVGPathTransitions &
+          SVGForcedAttrTransitions &
+          SVGTransitions &
+          VariableTransitions & {
+              default?: ValueTransition
+              layout?: ValueTransition
+          })
+
 export interface TransformProperties {
-    x?: string | number
-    y?: string | number
-    z?: string | number
-    translateX?: string | number
-    translateY?: string | number
-    translateZ?: string | number
-    rotate?: string | number
-    rotateX?: string | number
-    rotateY?: string | number
-    rotateZ?: string | number
-    scale?: string | number
-    scaleX?: string | number
-    scaleY?: string | number
-    scaleZ?: string | number
-    skew?: string | number
-    skewX?: string | number
-    skewY?: string | number
-    originX?: string | number
-    originY?: string | number
-    originZ?: string | number
-    perspective?: string | number
-    transformPerspective?: string | number
+    x?: AnyResolvedKeyframe
+    y?: AnyResolvedKeyframe
+    z?: AnyResolvedKeyframe
+    translateX?: AnyResolvedKeyframe
+    translateY?: AnyResolvedKeyframe
+    translateZ?: AnyResolvedKeyframe
+    rotate?: AnyResolvedKeyframe
+    rotateX?: AnyResolvedKeyframe
+    rotateY?: AnyResolvedKeyframe
+    rotateZ?: AnyResolvedKeyframe
+    scale?: AnyResolvedKeyframe
+    scaleX?: AnyResolvedKeyframe
+    scaleY?: AnyResolvedKeyframe
+    scaleZ?: AnyResolvedKeyframe
+    skew?: AnyResolvedKeyframe
+    skewX?: AnyResolvedKeyframe
+    skewY?: AnyResolvedKeyframe
+    originX?: AnyResolvedKeyframe
+    originY?: AnyResolvedKeyframe
+    originZ?: AnyResolvedKeyframe
+    perspective?: AnyResolvedKeyframe
+    transformPerspective?: AnyResolvedKeyframe
 }
 
 export interface SVGForcedAttrProperties {